pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "labeltool"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A Python-based toolkit for scientific image segmentation refinement."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "numpy",
    "tifffile",
    "PyQt5",
    "pyqtgraph",
    "spam-package",
    "pyvista",
    "matplotlib",
    "scikit-image",
    "scipy",
    "colorcet",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/labeltool"
"Bug Tracker" = "https://github.com/yourusername/labeltool/issues"

[project.scripts]
labeltool = "labeltool.gui:main"

[tool.setuptools.packages.find]
where = ["src"]
